requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
Flask-WTF==1.2.1
email-validator==1.1.3




